personaplex-server/requirements.txt
fastapi
uvicorn
pydantic
aiohttp
torch
transformers
accelerate
soundfile
librosa
python-multipart
python-dotenv
